Add files via upload
A Theme Switcher project. The fundamental goal of this project is to be able to switch between the themes available. After the user changes the theme, it should be back to the default theme if needed. JavaScript was used to add the functionality. Concepts such as EventListeners were implemented. HTML was used to add the basic structure to the web page.